Made AttributeBuilder service partial
Moved methods for each attribute type to separate files
