+++ MageNet.Services/AttributeRepository/AttributeBuilder/AttributeBuilder.cs
@@ -1,17 +1,15 @@
 ﻿using MageNet.Persistence;
 using MageNet.Persistence.Models.AbstractModels.ModelEnums;
-using MageNet.Persistence.Models.Attributes;
 using MageNetServices.AttributeRepository.DTO.Attributes;
-using Microsoft.EntityFrameworkCore;
 using Attribute = MageNet.Persistence.Models.Attributes.Attribute;
 
-namespace MageNetServices.AttributeRepository;
+namespace MageNetServices.AttributeRepository.AttributeBuilder;
 
-public class AttributeDataProvider : IAttributeDataProvider
+public partial class AttributeBuilder : IAttributeBuilder
 {
     private readonly MageNetDbContext _dbContext;
 
-    public AttributeDataProvider(MageNetDbContext dbContext)
+    public AttributeBuilder(MageNetDbContext dbContext)
     {
         _dbContext = dbContext;
     }
@@ -166,160 +164,5 @@ public class AttributeDataProvider : IAttributeDataProvider
             throw new ArgumentException(
                 $"The wrong attribute id was provided. There is no attribute with id '{attributeWithData.AttributeId}'");
         }
-    }
-
-
-    private PriceAttribute GetPriceAttributeData(Attribute attribute)
-    {
-        var priceAttributeData = _dbContext.PriceAttributes
-            .FirstOrDefault(x => x.AttributeId == attribute.AttributeId);
-
-        if (priceAttributeData != null)
-        {
-            return priceAttributeData;
-        }
-
-        else
-        {
-            throw new ArgumentException(
-                $"The wrong attribute id was provided. There is no price attribute with id '{attribute.AttributeId}'");
-        }
-    }
-
-    private PriceAttribute CreatePriceAttributeData(AttributeWithData attributeWithData, Attribute attribute)
-    {
-        return new PriceAttribute
-        {
-            Attribute = attribute,
-            DefaultValue = decimal.Parse(attributeWithData.DefaultLiteralValue)
-        };
-    }
-
-    private PriceAttribute UpdatePriceAttributeData(AttributeWithData attributeWithData, Attribute attribute,
-        out bool isAttributeTypeChanged)
-    {
-        var priceAttribute =
-            _dbContext.PriceAttributes.SingleOrDefault(x => x.AttributeId == attributeWithData.AttributeId);
-        if (priceAttribute != null)
-        {
-            isAttributeTypeChanged = false;
-            priceAttribute.DefaultValue = Decimal.Parse(attributeWithData.DefaultLiteralValue);
-        }
-
-        else
-        {
-            isAttributeTypeChanged = true;
-            priceAttribute = CreatePriceAttributeData(attributeWithData, attribute);
-        }
-
-        return priceAttribute;
-    }
-
-    private TextAttribute GetTextAttributeData(Attribute attribute)
-    {
-        var textAttributeData = _dbContext.TextAttributes
-            .FirstOrDefault(x => x.AttributeId == attribute.AttributeId);
-
-        if (textAttributeData != null)
-        {
-            return textAttributeData;
-        }
-
-        else
-        {
-            throw new ArgumentException(
-                $"The wrong attribute id was provided. There is no text attribute with id '{attribute.AttributeId}'"
-            );
-        }
-    }
-
-
-    private TextAttribute CreateTextAttributeData(AttributeWithData attributeWithData, Attribute attribute)
-    {
-        return new TextAttribute
-        {
-            Attribute = attribute,
-            DefaultValue = attributeWithData.DefaultLiteralValue
-        };
-    }
-
-    private TextAttribute UpdateTextAttributeData(AttributeWithData attributeWithData, Attribute attribute,
-        out bool isAttributeTypeChanged)
-    {
-        var textAttribute =
-            _dbContext.TextAttributes.SingleOrDefault(x => x.AttributeId == attributeWithData.AttributeId);
-        if (textAttribute != null)
-        {
-            isAttributeTypeChanged = false;
-            textAttribute.DefaultValue = attributeWithData.DefaultLiteralValue;
-        }
-
-        else
-        {
-            isAttributeTypeChanged = true;
-            textAttribute = CreateTextAttributeData(attributeWithData, attribute);
-        }
-
-        return textAttribute;
-    }
-
-    private SelectableAttribute GetSelectableAttributeData(Attribute attribute)
-    {
-        var selectableAttributeData = _dbContext.SelectableAttributes.Include(x => x.Values)
-            .FirstOrDefault(x => x.AttributeId == attribute.AttributeId);
-
-        if (selectableAttributeData != null)
-        {
-            return selectableAttributeData;
-        }
-
-        else
-        {
-            throw new ArgumentException(
-                $"The wrong attribute id was provided. There is no selectable attribute with id '{attribute.AttributeId}'");
-        }
-    }
-
-    private SelectableAttribute CreateSelectableAttributeData(AttributeWithData attributeWithData, Attribute attribute)
-    {
-        var newValuesSet = new List<SelectableAttributeValue>();
-
-        if (attributeWithData.SelectableOptions != null)
-        {
-            foreach (var value in attributeWithData.SelectableOptions)
-            {
-                value.AttributeId = attribute.AttributeId;
-                newValuesSet.Add(value);
-            }
-        }
-
-        return new SelectableAttribute
-        {
-            Attribute = attribute,
-            IsMultipleSelect = (bool)attributeWithData.IsMultipleSelect,
-            Values = newValuesSet
-        };
-    }
-
-    private SelectableAttribute UpdateSelectableAttributeData(AttributeWithData attributeWithData, Attribute attribute,
-        out bool isAttributeTypeChanged)
-    {
-        var selectableAttribute =
-            _dbContext.SelectableAttributes.SingleOrDefault(x => x.AttributeId == attributeWithData.AttributeId);
-
-        if (selectableAttribute != null)
-        {
-            isAttributeTypeChanged = false;
-            selectableAttribute.IsMultipleSelect = (bool)attributeWithData.IsMultipleSelect;
-            selectableAttribute.Values = attributeWithData.SelectableOptions;
-        }
-
-        else
-        {
-            isAttributeTypeChanged = true;
-            selectableAttribute = CreateSelectableAttributeData(attributeWithData, attribute);
-        }
-
-        return selectableAttribute;
     }
 }
